[FEAT] 미구현 기능 3가지 및 리팩토링
1. 강의 자료 페이지 삭제
2. 직급별 재생 속도
3. 수강 완료 시 포인트 확인
