+++ src/main/java/com/wanted/legendkim/domain/lecture/dto/LectureResponse.java
@@ -1,7 +1,7 @@
 package com.wanted.legendkim.domain.lecture.dto;
 // 강의 상세 조회 응답 DTO
 // Section 엔티티를 Lecture 의미로 가공해 내려줌
-import com.wanted.legendkim.domain.section.Section;
+import com.wanted.legendkim.domain.section.entity.Section;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
